+++ lib/src/eval/compiler/macros/branch.dart
@@ -1,5 +1,9 @@
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/compiler/context.dart';
+import 'package:dart_eval/src/eval/compiler/errors.dart';
 import 'package:dart_eval/src/eval/compiler/macros/macro.dart';
+import 'package:dart_eval/src/eval/compiler/model/label.dart';
 import 'package:dart_eval/src/eval/compiler/statement/statement.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
 import 'package:dart_eval/src/eval/runtime/runtime.dart';
@@ -9,11 +13,15 @@ StatementInfo macroBranch(
     {required MacroVariableClosure condition,
     required MacroStatementClosure thenBranch,
     MacroStatementClosure? elseBranch,
-    bool resolveStateToThen = false}) {
+    bool resolveStateToThen = false,
+    AstNode? source}) {
   ctx.beginAllocScope();
   ctx.enterTypeInferenceContext();
 
   final conditionResult = condition(ctx).unboxIfNeeded(ctx);
+  if (!conditionResult.type.isAssignableTo(ctx, CoreTypes.bool.ref(ctx))) {
+    throw CompileError("Conditions must have a static type of 'bool'", source);
+  }
 
   final rewriteCond = JumpIfFalse.make(conditionResult.scopeFrameOffset, -1);
   final rewritePos = ctx.pushOp(rewriteCond, JumpIfFalse.LEN);
@@ -22,7 +30,17 @@ StatementInfo macroBranch(
 
   ctx.inferTypes();
   ctx.beginAllocScope();
+  final label = CompilerLabel(LabelType.branch, -1, (_ctx) {
+    _ctx.endAllocScopeQuiet();
+    if (!resolveStateToThen) {
+      _ctx.resolveBranchStateDiscontinuity(_initialState);
+    }
+    _ctx.endAllocScopeQuiet();
+    return -1;
+  });
+  ctx.labels.add(label);
   final thenResult = thenBranch(ctx, expectedReturnType);
+  ctx.labels.removeLast();
   ctx.endAllocScope();
   ctx.uninferTypes();
 
@@ -42,7 +60,15 @@ StatementInfo macroBranch(
 
   if (elseBranch != null) {
     ctx.beginAllocScope();
+    final label = CompilerLabel(LabelType.branch, -1, (_ctx) {
+      ctx.endAllocScope();
+      ctx.resolveBranchStateDiscontinuity(_initialState);
+      ctx.endAllocScope();
+      return -1;
+    });
+    ctx.labels.add(label);
     final elseResult = elseBranch(ctx, expectedReturnType);
+    ctx.labels.removeLast();
     ctx.endAllocScope();
     ctx.resolveBranchStateDiscontinuity(_initialState);
     ctx.rewriteOp(rewriteOut!, JumpConstant.make(ctx.out.length), 0);