+++ test/collection_test.dart
@@ -171,5 +171,21 @@ void main() {
 
       expect(runtime.executeLib('package:eval_test/main.dart', 'main'), true);
     });
+
+    test('Add key to empty map', () {
+      final runtime = compiler.compileWriteAndLoad({
+        'eval_test': {
+          'main.dart': '''
+            bool main() {
+              final testMap = <String, String>{};
+              testMap['name'] = 'Jon';
+              return testMap.isNotEmpty;
+            }
+          '''
+        }
+      });
+
+      expect(runtime.executeLib('package:eval_test/main.dart', 'main'), true);
+    });
   });
 }